PullRequest abfrage
Funktioniert alles außer closedBy bzw review

+++ api/src/main/java/Default/Issue/IssueRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
 Repository for Issues identified by ID (long)
  */
-public interface IssueRepository extends JpaRepository<Issue, Long>{
+public interface IssueRepository extends JpaRepository<Issue, Integer>{
 }
